requirements.txt
# Core dependencies
torch==2.0.0
torchvision==0.15.0

# Image processing
opencv-python==4.3.0.36
opencv-contrib-python==4.3.0.36
Pillow>=9.0.0
scikit-image==0.20.0

# Scientific computing
numpy==1.24.4
scipy==1.9.1
matplotlib==3.7.1

# Visualization and monitoring
visdom==0.2.4
tensorboard
tqdm

# Configuration
pyyaml
jsonargparse

# Data handling
h5py

# Evaluation metrics
lpips
pytorch-fid

# Optional for advanced features
# wandb  # For experiment tracking
# albumentations  # For advanced data augmentation